requirements.txt
playwright==1.40.0
openai==1.12.0
python-dotenv==1.0.0
matplotlib==3.8.2
Pillow==10.2.0
pydantic==2.5.3
mcp==0.9.0
dashscope>=1.17.0

